Simplify VS Code CLI E2E helpers
Extract shared CLI helper utilities and add an ExTester-to-test-cli Create Workspace parity map for traceability.
